Move patch_merge to collection helpers
