+++ test/teiserver/tachyon_lobby/lobby_test.exs
@@ -10,7 +10,6 @@ defmodule Teiserver.TachyonLobby.LobbyTest do
   use Teiserver.DataCase
 
   import Teiserver.Support.Polling, only: [poll_until_some: 1, poll_until_nil: 1]
-  import LobbyProcess, only: [patch_merge: 2]
 
   @moduletag :tachyon
 
@@ -1958,39 +1957,6 @@ defmodule Teiserver.TachyonLobby.LobbyTest do
     end
   end
 
-  # again, this should probably be exatracted in a more general module
-  describe "patch merge" do
-    test "update a simple (non map) value" do
-      assert patch_merge(%{key: "s1"}, %{key: "s2"}) == %{key: "s2"}
-
-      result = patch_merge(%{"string-key" => "s1"}, %{"string-key" => "s2"})
-      assert result == %{"string-key" => "s2"}
-    end
-
-    test "can add new keys" do
-      result = patch_merge(%{key: "foo"}, %{other: 2})
-      assert result == %{key: "foo", other: 2}
-    end
-
-    test "can delete keys when value is nil" do
-      result = patch_merge(%{foo: "fooval", bar: "barkey"}, %{bar: nil})
-      assert result == %{foo: "fooval"}
-    end
-
-    test "set map as value when new key" do
-      result = patch_merge(%{}, %{foo: %{key: "val"}})
-      assert result == %{foo: %{key: "val"}}
-    end
-
-    test "can recursively update nested maps" do
-      result =
-        patch_merge(%{foo: %{key: "base-key", foo: "bar"}}, %{
-          foo: %{key: 2, foo: nil, bar: "updated"}
-        })
-
-      assert result == %{foo: %{key: 2, bar: "updated"}}
-    end
-  end
 
   defp mk_start_params(teams) do
     %LT.StartParams{
